+++ app/src/main/java/eu/faircode/email/ServiceSynchronize.java
@@ -182,7 +182,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
             "ssl_harden", "ssl_harden_strict", "cert_strict", "cert_transparency", "check_names", "bouncy_castle", "bc_fips", // force reconnect
             "experiments", "debug", "protocol", // force reconnect
             //"restart_interval", // force reconnect
-            "auth_plain", "auth_login", "auth_ntlm", "auth_sasl", "auth_apop", // force reconnect
+            "imap_compress", "auth_plain", "auth_login", "auth_ntlm", "auth_sasl", "auth_apop", // force reconnect
             "keep_alive_poll", "empty_pool", "idle_done", // force reconnect
             "exact_alarms" // force schedule
     ));
@@ -2034,7 +2034,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
                                     long start = new Date().getTime();
                                     try {
                                         wlMessage.acquire(Helper.WAKELOCK_MAX);
-                                        fetch(folder, ifolder, e.getMessages(), false, false, "added");
+                                        fetch(state, folder, ifolder, e.getMessages(), false, false, "added");
                                         Thread.sleep(FETCH_YIELD_DURATION);
                                     } catch (Throwable ex) {
                                         Log.e(folder.name, ex);
@@ -2054,7 +2054,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
                                     long start = new Date().getTime();
                                     try {
                                         wlMessage.acquire(Helper.WAKELOCK_MAX);
-                                        fetch(folder, ifolder, e.getMessages(), false, true, "removed");
+                                        fetch(state, folder, ifolder, e.getMessages(), false, true, "removed");
                                         Thread.sleep(FETCH_YIELD_DURATION);
                                     } catch (Throwable ex) {
                                         Log.e(folder.name, ex);
@@ -2080,7 +2080,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
                                     try {
                                         wlMessage.acquire(Helper.WAKELOCK_MAX);
                                         Message imessage = e.getMessage();
-                                        fetch(folder, ifolder, new Message[]{imessage}, true, false, "changed");
+                                        fetch(state, folder, ifolder, new Message[]{imessage}, true, false, "changed");
                                         Thread.sleep(FETCH_YIELD_DURATION);
                                     } catch (Throwable ex) {
                                         Log.e(folder.name, ex);
@@ -2939,7 +2939,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
         }
     }
 
-    private void fetch(EntityFolder folder, IMAPFolder ifolder, Message[] messages, boolean invalidate, boolean deleted, String reason) throws MessagingException {
+    private void fetch(Core.State state, EntityFolder folder, IMAPFolder ifolder, Message[] messages, boolean invalidate, boolean deleted, String reason) throws MessagingException {
         Log.i(folder.name + " " + messages.length + " messages " + reason);
 
         List<Long> uids = new ArrayList<>();
@@ -2949,6 +2949,13 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
                 uids.add(uid);
             } catch (MessageRemovedException ex) {
                 Log.w(ex);
+            } catch (Throwable ex) {
+                Log.w(ex);
+                if (ex instanceof FolderClosedException) {
+                    state.error(ex);
+                    return;
+                } else
+                    throw ex;
             }
 
         Log.i(folder.name + " messages " + reason + " uids=" + TextUtils.join(",", uids));
@@ -3498,7 +3505,7 @@ public class ServiceSynchronize extends ServiceBase implements SharedPreferences
         if (!found) {
             if (BuildConfig.DEBUG)
                 Log.i("@@@ not found");
-            return null;
+            return new long[]{0, 0}; // no day found
         }
 
         long start = calStart.getTimeInMillis();