+++ CHANGELOG.md
@@ -30,7 +30,26 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### Next version (Work in progress)
 
-### 1.2299 2025-09-30
+### 1.2302 2025-10-30
+
+* Fixed all reported bugs
+* Updated [translations](https://crowdin.com/project/open-source-email)
+* Updated [Public Suffix List](https://github.com/publicsuffix/list)
+
+### 1.2301 2025-10-23
+
+* Fixed all reported bugs
+* Updated [translations](https://crowdin.com/project/open-source-email)
+* Updated [Public Suffix List](https://github.com/publicsuffix/list)
+
+### 1.2300 2025-10-06
+
+* Fixed all reported bugs
+* Added collapsible attachments for the message editor
+* Updated [S/MIME root certificates](https://wiki.mozilla.org/CA/Included_Certificates)
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
+### 1.2299 2025-09-30 *
 
 * Fixed all reported bugs
 * Small improvements and minor bug fixes